% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catimp.R
\name{catImp}
\alias{catImp}
\title{Categorical data model imputation using log linear models}
\usage{
catImp(obsData, M = 10, pd = FALSE, type = 1, margins = NULL,
  steps = 100, rseed)
}
\arguments{
\item{obsData}{The data frame to be imputed. Variables must be coded taking positive
integer values only.}

\item{M}{Number of imputations to generate.}

\item{pd}{Specify whether to use posterior draws (\code{TRUE})
or not (\code{FALSE}).}

\item{type}{An integer specifying what type of log-linear model to impute using.
\code{type=1}, the default, allows for all two-way associations in the log-linear model.
\code{type=2} allows for all three-way associations (plus lower).
\code{type=3} fits a saturated model.}

\item{margins}{An optional argument that can be used instead of \code{type} to specify
the desired log-linear model.}

\item{steps}{If \code{pd} is \code{TRUE}, the \code{steps} argument specifies
how many MCMC iterations to perform.}

\item{rseed}{The value to set the \code{cat} package's random number seed to,
using the \code{rngseed} function of \code{cat}. This function must be called at least
once before imputing using \code{cat}. If the user wishes to set the seed using
\code{rngseed} before calling \code{catImp}, set \code{rseed=NULL}.}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
This function performs multiple imputation under a log-linear model
as described by Schafer (1997), using his \code{cat} package. By default
\code{catImp} will impute using a log-linear model allowing for all two-way
associations, but not higher order associations. This can be modified through
use of the \code{type} and \code{margins} arguments.
}
