% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catimp.R
\name{catSatImp}
\alias{catSatImp}
\title{Categorical data model imputation under saturated multinomial model}
\usage{
catSatImp(obsData, M = 10, pd = FALSE, steps = 100)
}
\arguments{
\item{obsData}{The data frame to be imputed.}

\item{M}{Number of imputations to generate.}

\item{pd}{Specify whether to use posterior draws (\code{TRUE})
or not (\code{FALSE}).}

\item{steps}{If \code{pd} is \code{TRUE}, the \code{steps} argument specifies
how many MCMC iterations to perform.}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
This function performs multiple imputation under a saturated multinomial model
for categorical data, as described by Schafer (1997). If the dataset has
more than a small number of variables, a saturated multinomial model may not
be estimable, in which case the user should consider a simpler log-linear
model.
}
\details{
The \code{rngseed} function must be called at least once before \code{catSatImp}
is called.
}
