% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miximp.R
\name{mixImp}
\alias{mixImp}
\title{Imputation for a mixture of continuous and categorical variables using
the general location model.}
\usage{
mixImp(obsData, nCat, M = 10, pd = FALSE, type = 1, margins = NULL,
  design = NULL, steps = 100, rseed)
}
\arguments{
\item{obsData}{The data frame to be imputed. The categorical variables must be
in the first \code{p} columns, and they must be coded using consecutive positive
integers.}

\item{M}{Number of imputations to generate.}

\item{pd}{Specify whether to use posterior draws (\code{TRUE})
or not (\code{FALSE}).}

\item{margins}{An optional margins argument for specifying a restricted log-linear
model for the categorical variables. If \code{margins} is specified, \code{design}
must also be specified.}

\item{design}{An optional argument for specifying the design matrix controlling the
dependence of the continuous variables on the categorical variables.  If
\code{design} is specified, \code{margins} must also be specified.}

\item{steps}{If \code{pd} is \code{TRUE}, the \code{steps} argument specifies
how many MCMC iterations to perform.}

\item{rseed}{The value to set the \code{mix} package's random number seed to,
using the \code{rngseed} function of \code{mix}. This function must be called at least
once before imputing using \code{mix}. If the user wishes to set the seed using
\code{rngseed} before calling \code{mixImp}, set \code{rseed=NULL}.}

\item{p}{The number of categorical variables in \code{obsData}.}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
This function performs multiple imputation under a general location model
as described by Schafer (1997), using his \code{mix} package.
}
