% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refBasedCts.r
\name{refBasedCts}
\alias{refBasedCts}
\title{Reference based imputation of repeated measures continuous data}
\usage{
refBasedCts(obsData, outcomeVarStem, nVisits, trtVar,
  baselineVars = NULL, baselineVisitInt = FALSE, type = "MAR",
  M = 5)
}
\arguments{
\item{obsData}{The data frame to be imputed.}

\item{outcomeVarStem}{String for stem of outcome variable name, e.g. y if y1, y2, y3 are the outcome columns}

\item{nVisits}{The integer number of visits (not including baseline)}

\item{trtVar}{The string variable name of the treatment variable}

\item{baselineVars}{A string or vector of strings specfying the baseline variables. Often this will include
the baseline measurement of the outcome}

\item{type}{A string specifying imputation type to use. Valid options are "MAR", "J2R"}

\item{M}{Number of imputations to generate.}

\item{baselinVisitInt}{TRUE/FALSE indicating whether to allow for interactions between each baseline variable
and visit}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
Performs multiple imputation of a repeatedly measured continuous endpoint
using reference based imputation as proposed by Carpenter et al. This approach
can be used for imputation of missing data in randomised clinical trials.
}
\details{
Due to uncongeniality between the imputation and analysis models, the resulting
imputed datasets should be analysed using the
\href{https://cran.r-project.org/package=bootImpute}{bootImpute} package.
Unlike most implementations of reference based imputation, this implementation
imputes conditional on the maximum likelihood estimates of the model parameters,
rather than a posterior draw. This is ok provided the aforementioned bootstrapping
approach is used for inference, rather than Rubin's rules.

Intermediate missing values are imputed assuming MAR, based on the mixed model fit
to that patient's treatment arm. Monotone missing values are imputed using the specified
imputation type.

Baseline covariates must be numeric variables. If you have factor variables you must
code these into suitable dummy indicators and pass these to the function.

THINGS TO DO:
1) option for INTERACTIONS BETWEEN BASELINE AND VISIT
2) implement copy reference method
}
\examples{
#simulate a dataset with one partially observed (conditionally) normal variable
set.seed(1234)
n <- 100
x <- rnorm(n)
y <- x+rnorm(n)
x[runif(n)<0.25] <- NA
temp <- data.frame(x=x,y=y)

#impute using normImp
imps <- normUniImp(temp, y~x, M=10, pd=FALSE)
}
\references{
von Hippel P.T. (2018) Maximum likelihood multiple imputation: faster,
more efficient imputation without posterior draws. \href{https://arxiv.org/abs/1210.0870v9}{arXiv:1210.0870v9}.
}
