% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.r
\name{scoreBased}
\alias{scoreBased}
\title{Score based variance estimation}
\usage{
scoreBased(imps, analysisFun, pd = NULL, dfComplete = NULL, ...)
}
\arguments{
\item{imps}{A list of imputed datasets produced by one of the imputation functions
in \code{mlmi} or another package.}

\item{analysisFun}{A function to analyse the imputed datasets that when applied to
a dataset returns a list containing a vector \code{est}.}

\item{dfComplete}{The complete data degrees of freedom. If \code{analysisFun} returns a vector
of parameter estimates, \code{dfComplete} should be a vector of the same length. If not
specified, it is assumed that the complete data degrees of freedom is effectively infinite (1e+05).}

\item{...}{Other parameters that are to be passed through to \code{analysisFun}.}

\item{scoreFun}{A function whose first argument is a dataset and whose second argument is
a vector of parameter values. It should return a matrix of subject level scores \code{scores}
evaluated at the parameter value passed to it.}
}
\value{
A list containing the overall parameter estimates, its corresponding covariance matrix, and
degrees of freedom for each parameter.
}
\description{
This function implements the score based variance estimation approach described by von Hippel
and Bartlett, based on Wang and Robins (1998).
}
\examples{
#simulate a partially observed dataset
set.seed(1234)
n <- 100
x <- rnorm(n)
y <- x+rnorm(n)
y[1:50] <- NA
temp <- data.frame(x,y)

#impute using normImp
imps <- normImp(temp, M=100, pd=TRUE, rseed=4423)

#analyse
analysisFun <- function(inputData) {
  mod <- lm(y~x, data=inputData)
  list(est=coef(mod), var=vcov(mod))
}
withinBetween(imps,analysisFun, dfComplete=c(n-2,n-2))
}
\references{
von Hippel PT, Bartlett JW. Maximum likelihood multiple imputation: can multiple imputation
work without posterior draws? under review
}
