% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.r
\name{withinBetween}
\alias{withinBetween}
\title{Within between variance estimation}
\usage{
withinBetween(imps, analysisfun, dfComplete = 1e+05, ...)
}
\arguments{
\item{imps}{A list of imputed datasets.}

\item{analysisfun}{A function to analyse the imputed datasets.}

\item{dfComplete}{The complete data degrees of freedom.}

\item{...}{Other parameters that are to be passed through to \code{analysisfun}.}
}
\value{
A list of imputed datasets.
}
\description{
Within between variance estimation
}
