% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normimp.R
\name{normimp}
\alias{normimp}
\title{Multivariate normal model imputation}
\usage{
normimp(obsData, M = 10, pd = FALSE, steps = 100)
}
\arguments{
\item{obsData}{The data frame to be imputed.}

\item{M}{Number of imputations to generate.}

\item{pd}{Specify whether to use posterior draws (\code{TRUE})
or not (\code{FALSE}).}

\item{steps}{If \code{pd} is \code{TRUE}, the \code{steps} argument specifies
how many MCMC iterations to perform.}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
Multivariate normal model imputation
}
