% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmi.r
\name{normUniImp}
\alias{normUniImp}
\title{Normal regression imputation of a single variable}
\usage{
normUniImp(obsData, impFormula, M = 5, pd = FALSE)
}
\arguments{
\item{obsData}{The data frame to be imputed.}

\item{impFormula}{The linear model formula.}

\item{M}{Number of imputations to generate.}

\item{pd}{Specify whether to use posterior draws (\code{TRUE})
or not (\code{FALSE}).}
}
\value{
A list of imputed datasets, or if \code{M=1}, just the imputed data frame.
}
\description{
Performs multiple imputation of a single continuous variable using a normal
linear regression model. The covariates in the imputation model must be fully
observed. By default \code{normUniImp} imputes every dataset using the
maximum likelihood estimates of the imputation model parameters, which here
coincides with the OLS estimates. If \code{pd=TRUE} is specified, it instead
performs posterior draw Bayesian imputation.
}
